Fix crash when speaking to some NPC. Fix event system collection modified exception. Fix double cast? Fix trait packet parsing

+++ Source/Game/Entities/Player/Player.cs
@@ -5045,7 +5045,7 @@ namespace Game.Entities
 
             PushQuests();
 
-            Global.ScriptMgr.ForEach<IPlayerOnLevelChanged>(p => p.OnLevelChanged(this, (byte)oldLevel));
+            Global.ScriptMgr.ForEach<IPlayerOnLevelChanged>(p => p.OnLevelChanged(this, oldLevel));
         }
 
         public bool CanParry()
